Changes due to pull request
Task: #24
