+++ de.tu_bs.cs.isf.e4cf.parts.project_explorer/src/de/tu_bs/cs/isf/e4cf/parts/project_explorer/CustomTreeCell.java
@@ -8,28 +8,18 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
 
 import de.tu_bs.cs.isf.e4cf.core.file_structure.FileTreeElement;
 import de.tu_bs.cs.isf.e4cf.core.file_structure.WorkspaceFileSystem;
 import de.tu_bs.cs.isf.e4cf.core.file_structure.components.Directory;
 import de.tu_bs.cs.isf.e4cf.core.file_structure.util.FileHandlingUtility;
-import de.tu_bs.cs.isf.e4cf.core.stringtable.E4CStringTable;
 import de.tu_bs.cs.isf.e4cf.core.util.RCPMessageProvider;
-import de.tu_bs.cs.isf.e4cf.core.util.ServiceContainer;
-import de.tu_bs.cs.isf.e4cf.parts.project_explorer.interfaces.IProjectExplorerExtension;
-import de.tu_bs.cs.isf.e4cf.parts.project_explorer.stringtable.FileTable;
-import javafx.embed.swt.SWTFXUtils;
-import javafx.event.EventHandler;
-import javafx.scene.Node;
 import javafx.scene.control.TextField;
 import javafx.scene.control.TreeItem;
 import javafx.scene.control.cell.TextFieldTreeCell;
-import org.eclipse.swt.graphics.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.image.WritableImage;
 import javafx.scene.input.DragEvent;
 import javafx.scene.input.Dragboard;
+import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.TransferMode;
 
@@ -39,54 +29,42 @@ import javafx.scene.input.TransferMode;
 public class CustomTreeCell extends TextFieldTreeCell<FileTreeElement> {
 
 	private TextField editTextField;
-	private Map<String, IProjectExplorerExtension> fileExtensions;
-	private ServiceContainer services;
-	
-	public CustomTreeCell(WorkspaceFileSystem workspaceFileSystem, ServiceContainer service, Map<String, IProjectExplorerExtension> fileExtensions) {
-		super();
-		
-		this.services = service;
-		this.fileExtensions = fileExtensions;
+	private FileImageProvider fileImageProvider;
 
-		setOnDragOver(new EventHandler<DragEvent>() {
-			@Override
-			public void handle(DragEvent event) {
-				event.acceptTransferModes(TransferMode.COPY);
-			}
-		});
+	public CustomTreeCell(WorkspaceFileSystem workspaceFileSystem, FileImageProvider fileImageProvider) {
+		this.fileImageProvider = fileImageProvider;
 
-		setOnDragDropped(new EventHandler<DragEvent>() {
+		setOnDragOver((DragEvent event) -> event.acceptTransferModes(TransferMode.COPY));
 
-			@Override
-			public void handle(DragEvent event) {
-				TreeItem<FileTreeElement> currentItem = getTreeItem();
-				Directory directory = (Directory) currentItem.getValue();
+		setOnDragDropped((DragEvent event) -> {
+			TreeItem<FileTreeElement> currentItem = getTreeItem();
+			Directory directory = (Directory) currentItem.getValue();
 
-				final Dragboard db = event.getDragboard();
-				boolean success = false;
-				if (db.hasFiles()) {
-					List<java.io.File> files = db.getFiles();
-					for (java.io.File file : files) {
-						try {
-							workspaceFileSystem.copy(Paths.get(file.getAbsolutePath()),
-									Paths.get(directory.getAbsolutePath()));
-							success = true;
-						} catch (FileAlreadyExistsException e) {
-							RCPMessageProvider.errorMessage("File already exsits.",
-									"A file with the name " + file + " exists.");
-						} catch (IOException e) {
-							e.printStackTrace();
-						}
+			final Dragboard db = event.getDragboard();
+			boolean success = false;
+			if (db.hasFiles()) {
+				List<java.io.File> files = db.getFiles();
+				for (java.io.File file : files) {
+					try {
+						workspaceFileSystem.copy(Paths.get(file.getAbsolutePath()),
+								Paths.get(directory.getAbsolutePath()));
+						success = true;
+					} catch (FileAlreadyExistsException e) {
+						RCPMessageProvider.errorMessage("File already exsits.",
+								"A file with the name " + file + " exists.");
+					} catch (IOException e) {
+						e.printStackTrace();
 					}
 				}
-				event.setDropCompleted(success);
-				event.consume();
 			}
+			event.setDropCompleted(success);
+			event.consume();
 		});
 	}
-	
+
 	/**
-	 * Will be invoked by pressing F2, Enter, double click (default behavior by JavaFX)
+	 * Will be invoked by pressing F2, Enter, double click (default behavior by
+	 * JavaFX)
 	 */
 	@Override
 	public void startEdit() {
@@ -103,17 +81,16 @@ public class CustomTreeCell extends TextFieldTreeCell<FileTreeElement> {
 	public void cancelEdit() {
 		super.cancelEdit();
 		setText(getItem().toString());
-		setGraphic(getImage(getItem()));
+		setGraphic(fileImageProvider.getImage(getItem()));
 	}
-
 
 	@Override
 	public void updateItem(FileTreeElement item, boolean empty) {
 		super.updateItem(item, empty);
 
 		if (!empty) {
-			setText(item.toString());			
-			setGraphic(getImage(item));
+			setText(item.toString());
+			setGraphic(fileImageProvider.getImage(item));
 		} else {
 			setText("");
 			setGraphic(null);
@@ -121,98 +98,46 @@ public class CustomTreeCell extends TextFieldTreeCell<FileTreeElement> {
 	}
 
 	private void renameItem() {
-        String newText = editTextField.getText();
-        Path source = FileHandlingUtility.getPath(getItem());
-        Path target = source.getParent().resolve(newText);
+		String newText = editTextField.getText();
+		Path source = FileHandlingUtility.getPath(getItem());
+		Path target = source.getParent().resolve(newText);
 
+		File sourceFile = source.toFile();
 
-        File sourceFile = source.toFile();
+		if (!sourceFile.isDirectory() || (sourceFile.isDirectory() && sourceFile.list().length == 0)) {
+			try {
+				Files.move(source, target);
+			} catch (IOException e) {
+				RCPMessageProvider.errorMessage("Rename File", e.getMessage());
+			}
+		} else {
+			// Traverse the file tree and copy each file/directory.
+			try {
 
-        if (!sourceFile.isDirectory() || (sourceFile.isDirectory() && sourceFile.list().length == 0)) {
-            try {
-                Files.move(source, target);
-            } catch (IOException e) {
-                RCPMessageProvider.errorMessage("Rename File", e.getMessage());            }
-        } else {
-            // Traverse the file tree and copy each file/directory.
-            try {
+				Files.walk(source).forEach(sourcePath -> {
+					Path targetPath = target.resolve(source.relativize(sourcePath));
+					try {
+						Files.copy(sourcePath, targetPath);
+					} catch (IOException e) {
+						e.printStackTrace();
+					}
+				});
 
-                Files.walk(source)
-                        .forEach(sourcePath -> {
-                                Path targetPath = target.resolve(source.relativize(sourcePath));
-                                try {
-                                    Files.copy(sourcePath, targetPath);
-                                } catch (IOException e) {
-                                    e.printStackTrace();
-                                }
-                        });
+				Files.walk(source).sorted(Comparator.reverseOrder()).map(Path::toFile).forEach(File::delete);
 
-
-                Files.walk(source)
-                  .sorted(Comparator.reverseOrder())
-                  .map(Path::toFile)
-                  .forEach(File::delete);
-
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-    }
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		}
+	}
 
 	private void setupEditTextField() {
 		editTextField = new TextField(getItem().toString());
 		editTextField.selectAll();
-		editTextField.setOnKeyReleased(new EventHandler<KeyEvent>() {
-
-			@Override
-			public void handle(KeyEvent event) {
-				switch (event.getCode()) {
-				case ENTER:
-					renameItem();
-					break;
-				default:
-					break;
-				}
+		editTextField.setOnKeyReleased((KeyEvent event) -> {
+			if (event.getCode() == KeyCode.ENTER) {
+				renameItem();
 			}
 		});
 	}
-
-
-	/**
-	 * Returns an appropriate image for a given tree element
-	 * @param element tree element
-	 * @return an image
-	 * TODO: Remove this method from the controller
-	 */
-	public Node getImage(Object element) {
-		Image image = null;
-
-		if (element instanceof FileTreeElement) {
-			FileTreeElement fileElement = (FileTreeElement) element;
-			if (fileElement.getParent() == null) {
-				return null;
-			} else if (services.workspaceFileSystem.isProject(fileElement)) {
-				image = services.imageService.getImage(null, FileTable.PROJECT_PNG);
-			} else if (fileElement.isDirectory()) {
-				image = services.imageService.getImage(null, FileTable.FOLDER_PNG);
-			} else {
-				String fileExtension = fileElement.getExtension();
-				// load extended file icons
-				if (fileExtensions.containsKey(fileExtension)) {
-					image = fileExtensions.get(fileExtension).getIcon(services.imageService);
-				} else if (fileExtension.equals(E4CStringTable.FILE_ENDING_XML)) {
-					image = services.imageService.getImage(null, FileTable.XML_PNG);
-				} else {
-					// default file icon
-					image = services.imageService.getImage(null, FileTable.FILE_PNG);
-				}
-			}
-		}
-
-		WritableImage fxImage = SWTFXUtils.toFXImage(image.getImageData(), null);
-
-		return new ImageView(fxImage);
-	}
-
 }
