+++ de.tu_bs.cs.isf.e4cf.parts.project_explorer/src/de/tu_bs/cs/isf/e4cf/parts/project_explorer/controller/ProjectExplorerViewController.java
@@ -31,6 +31,7 @@ import de.tu_bs.cs.isf.e4cf.core.util.RCPContentProvider;
 import de.tu_bs.cs.isf.e4cf.core.util.ServiceContainer;
 import de.tu_bs.cs.isf.e4cf.core.util.extension_points.ExtensionAttrUtil;
 import de.tu_bs.cs.isf.e4cf.parts.project_explorer.CustomTreeCell;
+import de.tu_bs.cs.isf.e4cf.parts.project_explorer.FileImageProvider;
 import de.tu_bs.cs.isf.e4cf.parts.project_explorer.interfaces.IProjectExplorerExtension;
 import de.tu_bs.cs.isf.e4cf.parts.project_explorer.interfaces.WorkspaceStructureTemplate;
 import de.tu_bs.cs.isf.e4cf.parts.project_explorer.listeners.OpenFileListener;
@@ -103,6 +104,8 @@ public class ProjectExplorerViewController {
 		canvas.getScene().addEventFilter(MouseEvent.MOUSE_CLICKED,
 				new OpenFileListener(context, fileExtensions, services));
 
+		FileImageProvider fileImageProvider = new FileImageProvider(services, fileExtensions);
+
 		// Cell factory for custom tree cells
 		projectTree.setEditable(true);
 
@@ -110,7 +113,7 @@ public class ProjectExplorerViewController {
 
 			@Override
 			public TreeCell<FileTreeElement> call(TreeView<FileTreeElement> param) {
-				TreeCell<FileTreeElement> treeCell = new CustomTreeCell(fileSystem, services, fileExtensions);
+				TreeCell<FileTreeElement> treeCell = new CustomTreeCell(fileSystem, fileImageProvider);
 				return treeCell;
 			}
 		});
@@ -228,17 +231,15 @@ public class ProjectExplorerViewController {
 		projectTree.setShowRoot(false);
 		projectTree.getSelectionModel().selectedItemProperty().addListener(changeListener);
 	}
-	
+
 	/**
-	 * Trigger Rename
+	 * Subscribing on the rename event
 	 */
 	@Inject
 	@Optional
 	public void rename(@UIEventTopic(E4CEventTable.EVENT_RENAME_PROJECT_EXPLORER_ITEM) Object o) {
 		projectTree.edit(projectTree.getSelectionModel().getSelectedItem());
 	}
-
-
 
 	/**
 	 * Traverse the filesystem tree via dfs to save the old state of each node
